+++ cli/subcommands/src/cargo_hax.rs
@@ -149,7 +149,12 @@ use hax_types::diagnostics::report::ReportCtx;
 
 #[extension_traits::extension(trait ExtHaxMessage)]
 impl HaxMessage {
-    fn report(self, message_format: MessageFormat, rctx: Option<&mut ReportCtx>) {
+    fn report(self, message_format: MessageFormat, mut rctx: Option<&mut ReportCtx>) {
+        if let (Some(r), HaxMessage::Diagnostic { diagnostic, .. }) = (rctx.as_mut(), &self)
+            && r.seen_already(diagnostic.clone())
+        {
+            return;
+        }
         match message_format {
             MessageFormat::Json => println!("{}", serde_json::to_string(&self).unwrap()),
             MessageFormat::Human => self.report_styled(rctx),